add domain restriction to public view
